Extracted out some helpers from RequestCallerTrait to another trait
